hapus

+++ admin/hapus.php
@@ -9,6 +9,9 @@ $res = mysqli_fetch_array($q);
 
 // menghapus data dari database dan menghapus file
 mysqli_query($koneksi,"delete from surat_mhs where Id='$id'");
+if(file_exists("../uploads/".$res['file']))
+{
 unlink("../uploads/".$res['file']);
+}
 // mengalihkan halaman kembali ke index.php
 header("location:suratku.php");
